Refactor VLA action entrypoints
