+++ docs_new/cookbook/vla/OpenPI/Pi0.5.mdx
@@ -409,7 +409,7 @@ payload = {
 }
 ```
 
-For OpenPI websocket clients, include equivalent `enable_pi_prefix_cache` and `enable_pi_cuda_graph` fields in each raw msgpack observation if you need per-request compatibility controls.
+For OpenPI websocket clients, include equivalent `enable_prefix_cache` and `enable_cuda_graph` fields in each raw msgpack observation if you need per-request compatibility controls.
 
 ### 6.4 Deployment Choices
